Scheduling Refactor and Database Appliance Scheduling
Story #135

By default, the driver now selects a database appliance owned by the current tenant if possible.
If the driver is initialized with a plumbing_hooks_class it uses the plumbing hooks' select_device.

Refactor of inventory and hooks to enable custom scheduling.
Plumbing Hooks has been split into two interfaces, network hooks and scheduling hooks.
Scheduling hooks returns an iterable of devices on which a new object can be created.

Scheduling hooks can return devices that aren't in the config or in the inventory.
The inital use of this is to select a database appliance owned by the tenant.
If the device returned from scheduling hooks isn't in the appliance inventory, it's added as an
a10 appliance db. Only the required configuration can be saved in a10 appliance db.
TODO: We need to add explicit support for each of our configuration options to the database model, Inventory device_appliance.

We'll likely change the scheduling hooks interface somewhat before reintegrating orchestration.

----

* simplified inventory mock for unit tests
* Expose old plumbing hooks as _plumbing_hooks so it's available to the migrations
* Tests for inventory/scheduling hooks refactor
* a10_context/inventory interaction tests