A10 appliance extension protocol and port
Support protocol and port optional device configuration in the a10 appliance extension.
Uses default protocol (https) and port (443 for https, 80 for http) if they aren't
specified.

ATTN
----

As a result of changing the migration that creates the a10_appliance_db table, this
change requires downgrading and then re-upgrading the database migrations. This won't
affect customers and reduces the number of migrations customers will need to run.
Adding or altering a NOT NULL integer column didn't work with sqlite for testing;
we can solve that problem later if we need to.

    $ a10-neutron-lbaas-db-manage downgrade 28a984ff83e1
    $ a10-neutron-lbaas-db-manage install

***

Required for story #135

***

* Exposed A10Config mechanism to populate device config defaults
* Apply device config defaults to devices being saved from the scheduler
* Apply device config defaults to devices being saved from the a10 appliance extension
* Added protocol and port to A10ApplianceDB model, migration
* Apply device config defaults to devices being read from a10_appliance_db.
  This is a temporary change until we support all of the configuration options.
* Template resource attribute map to use neutron api attributes